Change HoveredItem to uint

+++ Dalamud/Plugin/Services/IGameGui.cs
@@ -3,6 +3,7 @@ using System.Numerics;
 using Dalamud.Game.Gui;
 using Dalamud.Game.NativeWrapper;
 using Dalamud.Game.Text.SeStringHandling.Payloads;
+using Dalamud.Utility;
 
 namespace Dalamud.Plugin.Services;
 
@@ -17,9 +18,10 @@ public unsafe interface IGameGui : IDalamudService
     event Action<bool> UiHideToggled;
 
     /// <summary>
-    /// Event that is fired when the currently hovered item changes.
+    /// Event that is fired when the currently hovered item changes. 0 when no item is hovered.<br/>
+    /// Use <see cref="ItemUtil.GetBaseId"/> to get the base id and classification.
     /// </summary>
-    event Action<ulong> HoveredItemChanged;
+    event Action<uint> HoveredItemChanged;
 
     /// <summary>
     /// Event that is fired when the currently hovered action changes.
@@ -38,10 +40,10 @@ public unsafe interface IGameGui : IDalamudService
     bool GameUiHidden { get; }
 
     /// <summary>
-    /// Gets or sets the item ID that is currently hovered by the player. 0 when no item is hovered.
-    /// If > 1.000.000, subtract 1.000.000 and treat it as HQ.
+    /// Gets or sets the item ID that is currently hovered by the player. 0 when no item is hovered.<br/>
+    /// Use <see cref="ItemUtil.GetBaseId"/> to get the base id and classification.
     /// </summary>
-    ulong HoveredItem { get; set; }
+    uint HoveredItem { get; set; }
 
     /// <summary>
     /// Gets the action ID that is current hovered by the player. 0 when no action is hovered.
